django_frtb_backend/requirements.txt
django
django-cors-headers
django-extensions
djangorestframework
